+++ python/tests/test_api_consistency.py
@@ -88,6 +88,7 @@ EXCLUDED_TESTS_FILENAMES = {
     "async_only": [
         "test_deprecation_warnings.py",
         "test_client_side_cache.py",
+        "test_fork_safety.py",
     ],
     "sync_only": [
         "test_sync_client_side_cache.py",